Remove ms_full_const_get from MeiliSearch::ClassMethods
Since the ms_full_const_get is almost the same of defined in the rails
std lib (active_support), there is no value in keeping this copy.

Another problem was related to the implementation is about the system was finding
initially the meilisearch class not the app one, because the impl doesn't ensure the
entire ancestors chain.