Add a "submodule" called Rails
We are not supposed to use the same name definition to our ruby gem, we need to separate the contexts between them.
This will avoid potential issues in the future, for example, when I need to access something directly from the ruby gem
I could have some problem, or when I need to define the same constant in both places (like VERSION).

- [https://github.com/algolia/algoliasearch-client-ruby/blob/master/lib/algolia.rb](https://github.com/algolia/algoliasearch-client-ruby/blob/master/lib/algolia.rb)
    - algolia and algoliasearch (rails)
- [https://github.com/getsentry/sentry-ruby/tree/master/sentry-rails](https://github.com/getsentry/sentry-ruby/tree/master/sentry-rails)
    - sentry and sentry-rails
- [https://github.com/thoughtbot/factory_bot](https://github.com/thoughtbot/factory_bot)
    - factory_bot and factory_bot_rails
- [https://github.com/kaminari/kaminari](https://github.com/kaminari/kaminari)
    - kaminari and kaminari-core, kaminari-activerecord (and so on)
- [https://github.com/rubocop/rubocop](https://github.com/rubocop/rubocop)
    - rubocop and rubocop-rails, rubocop-rspec...