Revert to creating index asynchronously
Creating the index synchronously was required to get an up to version of
its settings, in order to compare them to the user configuration and
decide whether or not to update settings. Unfortunately this made
practically every operation have a synchronous component causing a lot
of problems down the line.
This commit reverts to using create_index asynchronously to avoid said
issues, and represents the settings of a nonexistent index as an empty
hash, which will queue an update_settings if the user has specified any
settings at all.