Fix will_paginate crash when Meilisearch disabled
Using Meilisearch::Rails.active? as an alternative to checking that the
objects being replaced are actually NullObject.

For an alternative solution see https://github.com/meilisearch/meilisearch-rails/pull/383/

The advantage of this approach is that NullObject remains opaque and
requires no special treatment, and the logic is much easier to follow
(in my opinion).