Add rspec status persistence file
This allows rspec to "remember" the last test run, allowing for commands
like `rspec --only-failures`, which should improve developer UX.

Picked `spec/tmp/examples.txt` since it's idiomatic to call this file
examples.txt and `spec/tmp` was already in the gitignore.

It's a small thing but I don't think there's an argument against having
it.